+++ src/codegen/function_call.rs
@@ -4,6 +4,8 @@ use super::Transpilable;
 
 impl Transpilable for FunctionCall {
     fn transpile(&self) -> String {
-        format!("{}()", self.function.transpile())
+        let parameters = &self.arguments.arguments.iter().map(|expr| expr.transpile()).collect::<Vec<_>>().join(", ");
+
+        format!("{}({})", self.function.transpile(), parameters)
     }
 }