+++ src/syntax/ast/functions.rs
@@ -3,6 +3,7 @@ use super::Expression;
 #[derive(Debug)]
 pub struct FunctionCall {
     pub function: Box<Expression>,
+    pub arguments: Box<ArgumentsList>,
 }
 
 #[derive(Debug)]